+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # Changelog
 
-- fix: missing files in packaged app
+- new: app throws warning when older matlab release is used than it was created with.